Make a pass separating atomic.numbers from numeric
