+++ mathics/builtin/drawing/plot.py
@@ -13,7 +13,7 @@ import palettable
 
 
 from mathics.core.evaluators import apply_N
-from mathics.builtin.atomic.numbers import chop
+from mathics.builtin.numeric import chop
 from mathics.builtin.base import Builtin
 from mathics.builtin.graphics import Graphics
 from mathics.builtin.drawing.graphics3d import Graphics3D
