+++ mathics/builtin/numbers/integer.py
@@ -93,86 +93,6 @@ class Ceiling(SympyFunction):
         return from_sympy(sympy.ceiling(x))
 
 
-class IntegerLength(Builtin):
-    """
-    <dl>
-    <dt>'IntegerLength[$x$]'
-        <dd>gives the number of digits in the base-10 representation of $x$.
-    <dt>'IntegerLength[$x$, $b$]'
-        <dd>gives the number of base-$b$ digits in $x$.
-    </dl>
-
-    >> IntegerLength[123456]
-     = 6
-    >> IntegerLength[10^10000]
-     = 10001
-    >> IntegerLength[-10^1000]
-     = 1001
-    'IntegerLength' with base 2:
-    >> IntegerLength[8, 2]
-     = 4
-    Check that 'IntegerLength' is correct for the first 100 powers of 10:
-    >> IntegerLength /@ (10 ^ Range[100]) == Range[2, 101]
-     = True
-    The base must be greater than 1:
-    >> IntegerLength[3, -2]
-     : Base -2 is not an integer greater than 1.
-     = IntegerLength[3, -2]
-
-    '0' is a special case:
-    >> IntegerLength[0]
-     = 0
-
-    #> IntegerLength /@ (10 ^ Range[100] - 1) == Range[1, 100]
-     = True
-    """
-
-    attributes = listable | protected
-
-    rules = {
-        "IntegerLength[n_]": "IntegerLength[n, 10]",
-    }
-
-    messages = {
-        "base": "Base `1` is not an integer greater than 1.",
-    }
-
-    def apply(self, n, b, evaluation):
-        "IntegerLength[n_, b_]"
-
-        n, b = n.get_int_value(), b.get_int_value()
-        if n is None or b is None:
-            evaluation.message("IntegerLength", "int")
-            return
-        if b <= 1:
-            evaluation.message("IntegerLength", "base", b)
-            return
-
-        if n == 0:
-            # special case
-            return Integer0
-
-        n = abs(n)
-
-        # O(log(digits))
-
-        # find bounds
-        j = 1
-        while b ** j <= n:
-            j *= 2
-        i = j // 2
-
-        # bisection
-        while i + 1 < j:
-            # assert b ** i <= n <= b ** j
-            k = (i + j) // 2
-            if b ** k <= n:
-                i = k
-            else:
-                j = k
-        return Integer(j)
-
-
 class BitLength(Builtin):
     """
     <dl>
